Force Linear taper on fades if in middle of a fade in or out

+++ AudioKit/Common/Internals/CoreAudio/Apple Code/ParameterRamper.hpp
@@ -37,10 +37,9 @@ public:
 
     float getSkew() const;
 
-    void setOffset(float offset);
+    void setOffset(uint32_t offset);
 
-    float getOffset() const;
-
+    uint32_t getOffset() const;
 
     void setUIValue(float value);
